pyproject.toml
[tool.poetry]
name = "alchemist"
version = "0.1.0"
description = ""
authors = ["lokhiufung <lokhiufung123@gmail.com>"]
readme = "README.md"
packages = [
    { include = "ibapi", from = "externals"}
]

[tool.poetry.dependencies]
python = "^3.11"
rich = "^13.7.1"
schedule = "^1.2.2"
pyarrow = "^16.1.0"
pandas = "^2.2.2"
fastapi = "^0.111.1"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
dash-bootstrap-components = "^1.6.0"
dash = "^2.17.1"
ray = {extras = ["default"], version = "^2.40.0"}
pyzmq = "^26.2.0"
sqlalchemy = "^2.0.36"
psycopg2-binary = "^2.9.10"
pytest = "^8.3.4"
pytest-cov = "^6.0.0"


[tool.poetry.group.dev.dependencies]
sphinx = "^8.0.2"
sphinx-rtd-theme = "^3.0.2"
furo = "^2024.8.6"
sphinx-material = "^0.0.36"
coverage = "^7.6.9"
commitizen = "^4.1.0"
sphinxcontrib-plantuml = "^0.30"
pylint = "^3.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
